requirements.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
google-generativeai
sentence-transformers
faiss-cpu